requirements.txt
flask==2.3.3
flask-cors==4.0.0
word-search-generator==1.2.0
python-dotenv==1.0.0